Correct error checking mistakes
In the partial update function hoursEstimated and completed were checked
all the time when they should not have been. This was introduced in the
split between checking whether it was given and it's value. In this case
we need to check if it was given again when we want to check it's value.

Also in the createComment function a variable was left as 'title' when
it should have been 'name'.