Update after lecture4
This is the state of my code after lecture4. It sets up axios to use the
Marvel Developer API, adds a form and sets it up to query the API.

I missed the bit about updating the state with the search results and
the bit about adding thumbnails to the character list.