Cleanup warnings and simplify 404 component
This cleans up a lot of warnings:
    - not specifying radix with `parseInt`
    - unused variables/imports
    - missing 'key' parameter after `map`

It adds better checks for error state in the paged versions of
components so that if the API returns a list of zero resources we can
redirect to a 404. (for example, this will happen if the user requests
/pokemon/page/1000)

This also simplifies the 404 page to make it easier to display. I had
issues using the redirects and now just embed it in place of results
when necessary, it seems to respond better.