Add pubsub sender shim and mostly fill out server
This adds 'nrp-sender-shim.js' to handle most of the work of sending
messages to the worker. The server is updated to use this and routes are
filled out, although the PUT route is likely to need work, message
bodies will probably be reworked to deal with this. The worker is also
updated to deal with the names of the messages it is expected to catch.