Add worker routes
This adds the worker routes for POST, DELETE and PUT as well as fixes a
small error in the server regarding error returns that do not have
'errorCode' attributes.

There is currently an issue with the POST and PUT routes, I believe
regarding updating the data in place. Requests time out for some reason,
I think it happens when searching for IDs.