Fix POST and PUT routes
The issue was that JSON.parse() was being called on JSON objects,
causing silent (for some reason?) errors. Once this was fixed the routes
behaved as expected.

The PUT route is also updated to use the ID from the route and not in
the request body (any id in the request body is ignored).