Add tex source for Lab 8
This updates the gitignore for LaTeX building with latexmk and adds a
source for Lab 8. This is using my own 'homework' class which can be
found in my dotfiles, it may be a good idea to add a note about this or
a link to it or something before releasing.